+++ src/lib.rs
@@ -130,7 +130,7 @@ use embedded_graphics::{
 };
 use object_chain::{Chain, ChainElement, Link};
 pub use parser::{ChangeTextStyle, Token};
-pub use rendering::TextBoxProperties;
+pub use rendering::{cursor::Cursor, TextBoxProperties};
 
 /// A text box object.
 ///