+++ src/plugin/tail.rs
@@ -1,6 +1,9 @@
 //! Display the last lines of the text.
 
-use embedded_graphics::{prelude::PixelColor, text::renderer::CharacterStyle};
+use embedded_graphics::{
+    prelude::PixelColor,
+    text::renderer::{CharacterStyle, TextRenderer},
+};
 
 use crate::{plugin::Plugin, rendering::cursor::Cursor, TextBoxProperties};
 
@@ -12,10 +15,10 @@ use crate::{plugin::Plugin, rendering::cursor::Cursor, TextBoxProperties};
 pub struct Tail;
 
 impl<'a, C: PixelColor> Plugin<'a, C> for Tail {
-    fn on_start_render<S: CharacterStyle>(
+    fn on_start_render<S: CharacterStyle + TextRenderer>(
         &mut self,
         cursor: &mut Cursor,
-        props: TextBoxProperties<'_, S>,
+        props: &TextBoxProperties<'_, S>,
     ) {
         if props.text_height > props.box_height {
             let offset = props.box_height - props.text_height;